backend/requirements.txt
SQLAlchemy>=1.4
psycopg2-binary>=2.9
dotenv
sqlacodegen
flask
flasgger
PyJWT
flask_jwt_extended
marshmallow
flask-cors